+++ python/configcorr/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-from .detmax_max_ent_K_config import *
-from .max_ent_config_naive_gradient_descent import *
+from .max_ent_config_dmcc import * 
+from .max_ent_config_naive_gradient_descent import *
